Fix float to integer altimeter meassure

+++ ETAPA III/Control_Ground_Stations/GCS3D/datauav.cpp
@@ -26,7 +26,7 @@ double dataUAV::longitude()
 
 double dataUAV::alture()
 {
-    return _Alture;
+    return (int16_t)_Alture;
 }
 
 double dataUAV::airspeed()
